refactor: bearer token code in own file
