+++ main.go
@@ -11,7 +11,6 @@ import (
 	goruntime "runtime"
 	"strconv"
 
-	"github.com/go-openapi/runtime"
 	"github.com/go-openapi/strfmt"
 	mcp_golang "github.com/metoro-io/mcp-golang"
 	"github.com/metoro-io/mcp-golang/transport/stdio"
@@ -37,58 +36,6 @@ type McpResponseLinks struct {
 var NO_NEXT_PAGE = McpResponseLinks{
 	NextPage:    0,
 	HasNextPage: false,
-}
-
-type BearerTokenMgr struct {
-	BearerToken string
-	IsReadOnly  bool
-	authInfo    runtime.ClientAuthInfoWriter
-}
-
-func NewBearerTokenMgr() *BearerTokenMgr {
-	b := &BearerTokenMgr{
-		BearerToken: "",
-		IsReadOnly:  false,
-		authInfo:    nil,
-	}
-	bearerToken, _ := os.LookupEnv("VANTAGE_BEARER_TOKEN")
-	if bearerToken == "" {
-		log.Printf("VANTAGE_BEARER_TOKEN not found, please create a read-only Service Token or Personal Access Token at https://console.vantage.sh/settings/access_tokens")
-		return b
-	}
-	b.BearerToken = bearerToken
-	client := meClient.NewClientWithBearerToken("api.vantage.sh", "/v2", "https", bearerToken)
-	response, err := client.GetMe(meClient.NewGetMeParams(), b.AuthInfo())
-	if err != nil {
-		return b
-	}
-	payload := response.GetPayload()
-	if payload == nil {
-		return b
-	}
-
-	if len(payload.BearerToken.Scope) == 1 && payload.BearerToken.Scope[0] == "read" {
-		b.IsReadOnly = true
-	}
-	return b
-}
-
-func (b *BearerTokenMgr) AuthInfo() runtime.ClientAuthInfoWriter {
-	if b.authInfo != nil {
-		return b.authInfo
-	}
-	b.authInfo = runtime.ClientAuthInfoWriterFunc(func(req runtime.ClientRequest, reg strfmt.Registry) error {
-		if err := req.SetHeaderParam("Authorization", fmt.Sprintf("Bearer %s", b.BearerToken)); err != nil {
-			return err
-		}
-
-		if err := req.SetHeaderParam("User-Agent", fmt.Sprintf("vantage-mcp-server/%s", Version)); err != nil {
-			return err
-		}
-
-		return nil
-	})
-	return b.authInfo
 }
 
 // Given the "next" URL from the API response, this function will return a
@@ -133,13 +80,14 @@ func setupLogger() {
 	}
 	logFile, err := os.OpenFile(logFilename, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0666)
 	if err != nil {
-		panic(fmt.Sprintf("Failed to open log file %s: %+v", err))
+		panic(fmt.Sprintf("Failed to open log file: %+v", err))
 	}
 	log.SetOutput(logFile)
 }
 
-func registerVantageTool[P any](server *mcp_golang.Server, bearerToken BearerTokenMgr, name string, description string, givenHandler func(params P) (*mcp_golang.ToolResponse, error)) {
-	server.RegisterTool(name, description, func(params P) (*mcp_golang.ToolResponse, error) {
+// run code that is common to all tools
+func registerVantageTool[ParamType any](server *mcp_golang.Server, bearerToken BearerTokenMgr, name string, description string, givenHandler func(params ParamType) (*mcp_golang.ToolResponse, error)) {
+	server.RegisterTool(name, description, func(params ParamType) (*mcp_golang.ToolResponse, error) {
 		log.Printf("invoked - tool - %s %+v", name, params)
 		if bearerToken.BearerToken == "" {
 			return nil, fmt.Errorf("VANTAGE_BEARER_TOKEN not found, please create a read-only Service Token or Personal Access Token at https://console.vantage.sh/settings/access_tokens")
